feat: i made a function that makes button disabled depends on situation
i made a function in order to forbidden duplicates from my cart list.
