+++ src/App.css
@@ -37,3 +37,6 @@ textarea {
   cursor: pointer;
   transition: all 0.4s;
 }
+.disabled {
+  background-color: #ad4300;
+}
